requirements.txt
accelerate==1.4.0
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
airportsdata==20250224
alpaca_eval==0.6.6
annotated-types==0.7.0
anyio==4.8.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==3.0.0
attrs==25.1.0
beautifulsoup4==4.13.3
blake3==1.0.4
bleach==6.2.0
blessed==1.20.0
bokeh==3.6.3
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cloudpickle==3.1.1
colorcet==3.1.0
comm==0.2.2
compressed-tensors==0.9.1
contourpy==1.3.1
cucim-cu12==25.2.0
cuda-bindings==12.8.0
cuda-python==12.8.0
cudf-cu12==25.2.2
cugraph-cu12==25.2.0
cuml-cu12==25.2.1
cuproj-cu12==25.2.0
cupy-cuda12x==13.4.0
cuspatial-cu12==25.2.0
cuvs-cu12==25.2.1
cuxfilter-cu12==25.2.0
cycler==0.12.1
dask==2024.12.1
dask-cuda==25.2.0
dask-cudf-cu12==25.2.2
dask-expr==1.1.21
datasets==3.3.2
datashader==0.17.0
debugpy==1.8.13
decorator==5.2.1
deepspeed==0.15.4
defusedxml==0.7.1
depyf==0.18.0
dill==0.3.8
diskcache==5.6.3
distributed==2024.12.1
distributed-ucxx-cu12==0.42.0
distro==1.9.0
dnspython==2.7.0
docker-pycreds==0.4.0
einops==0.8.1
email_validator==2.2.0
evaluate==0.4.3
executing==2.2.0
fast_jl==0.1.3
fastapi==0.115.11
fastapi-cli==0.0.7
fastjsonschema==2.21.1
fastrlock==0.8.3
filelock==3.17.0
fire==0.7.0
flash_attn==2.7.4.post1
fonttools==4.56.0
fqdn==1.5.1
frozenlist==1.5.0
fsspec==2024.12.0
geopandas==1.0.1
gguf==0.10.0
gitdb==4.0.12
GitPython==3.1.44
google-auth==2.38.0
google-genai==1.9.0
gpustat==1.1.1
h11==0.14.0
hf_transfer==0.1.9
hjson==3.1.0
holoviews==1.20.1
httpcore==1.0.7
httptools==0.6.4
httpx==0.28.1
huggingface-hub==0.29.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
iniconfig==2.0.0
inputimeout==1.0.4
InstructorEmbedding
interegular==0.3.3
ipykernel==6.29.5
ipython==9.0.2
ipython_pygments_lexers==1.1.1
isoduration==20.11.0
jedi==0.19.2
Jinja2==3.1.6
jiter==0.8.2
joblib==1.4.2
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter-events==0.12.0
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyter_server==2.15.0
jupyter_server_proxy==4.4.0
jupyter_server_terminals==0.5.3
jupyterlab_pygments==0.3.0
kiwisolver==1.4.8
lark==1.2.2
lazy_loader==0.4
libcudf-cu12==25.2.2
libcugraph-cu12==25.2.0
libcuml-cu12==25.2.1
libcuspatial-cu12==25.2.0
libcuvs-cu12==25.2.1
libkvikio-cu12==25.2.1
libraft-cu12==25.2.0
libucx-cu12==1.18.0
libucxx-cu12==0.42.0
linkify-it-py==2.0.3
llvmlite==0.43.0
lm-format-enforcer==0.10.11
locket==1.0.0
Markdown==3.7
MarkupSafe==3.0.2
matplotlib==3.10.1
matplotlib-inline==0.1.7
mdit-py-plugins==0.4.2
mistral_common==1.5.3
mistune==3.1.2
mpmath==1.3.0
msgpack==1.1.0
msgspec==0.19.0
multidict==6.1.0
multipledispatch==1.0.0
multiprocess==0.70.16
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.4.2
ninja==1.11.1.3
nltk==3.9.1
numba==0.60.0
numba-cuda==0.2.0
numpy==1.26.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-ml-py==12.570.86
nvidia-nccl-cu12==2.21.5
nvidia-nvcomp-cu12==4.2.0.11
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
nvtx==0.2.11
nx-cugraph-cu12==25.2.0
openai==1.65.4
opencv-python-headless==4.11.0.86
outlines==0.1.11
outlines_core==0.1.26
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pandocfilters==1.5.1
panel==1.6.1
param==2.2.0
parso==0.8.4
partd==1.4.2
partial-json-parser==0.2.1.1.post5
patsy==1.0.1
peft==0.14.0
pexpect==4.9.0
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
prometheus-fastapi-instrumentator==7.0.2
prometheus_client==0.21.1
propcache==0.3.0
protobuf==5.29.3
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybind11==2.13.6
pycountry==24.6.1
pycparser==2.22
pyct==0.5.0
pydantic==2.10.6
pydantic_core==2.27.2
pylibcudf-cu12==25.2.2
pylibcugraph-cu12==25.2.0
pylibraft-cu12==25.2.0
pynvjitlink-cu12==0.5.2
pynvml==12.0.0
pyogrio==0.10.0
pyparsing==3.2.1
pyproj==3.7.1
pytest==8.3.5
python-dotenv==1.0.1
python-json-logger==3.3.0
python-multipart==0.0.20
pytz==2025.1
pyviz_comms==3.0.4
PyYAML==6.0.2
pyzmq==26.2.1
qpop==1.5
raft-dask-cu12==25.2.0
rapids-dask-dependency==25.2.0
ray==2.40.0
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich-toolkit==0.13.2
rmm-cu12==25.2.0
rpds-py==0.23.1
rsa==4.9
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scikit-learn-extra==0.3.0
scipy==1.15.2
seaborn==0.13.2
Send2Trash==1.8.3
sentence-transformers==3.4.1
sentencepiece==0.2.0
sentry-sdk==2.22.0
setproctitle==1.3.5
setuptools==75.8.2
shapely==2.0.7
simpervisor==1.0.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.6
stack-data==0.6.3
starlette==0.46.0
sympy==1.13.1
tabulate==0.9.0
tblib==3.0.0
termcolor==2.5.0
terminado==0.18.1
threadpoolctl==3.5.0
tifffile==2025.2.18
tiktoken==0.9.0
tinycss2==1.4.0
tokenizers==0.21.0
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.2
tqdm==4.67.1
traitlets==5.14.3
traker==0.3.2
transformers==4.49.0
treelite==4.4.1
triton==3.1.0
typer==0.15.2
typer-slim==0.15.2
types-python-dateutil==2.9.0.20241206
tzdata==2025.1
uc-micro-py==1.0.3
ucx-py-cu12==0.42.0
ucxx-cu12==0.42.0
uri-template==1.3.0
urllib3==2.3.0
uvicorn==0.34.0
uvloop==0.21.0
vllm==0.7.3
wandb==0.19.8
watchfiles==1.0.4
webcolors==24.11.1
webencodings==0.5.1
websocket-client==1.8.0
websockets==15.0.1
wheel==0.45.1
xarray==2025.1.2
xformers==0.0.28.post3
xgrammar==0.1.11
xxhash==3.5.0
xyzservices==2025.1.0
yarl==1.18.3
zict==3.0.0
zipp==3.21.0
